#161 refactor: LivingBtn 기능 컴포넌트 내부에서 처리하도록 하기
